Rename context menu fallback scheduler
